requirements/development.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.4
django-autoslug==1.9.9
django-axes==6.1.1
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-dirtyfields==1.9.2
django-phonenumber-field==7.1.0
django-simple-history==3.3.0
django-versatileimagefield==2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
packaging==23.2
phonenumbers==8.13.19
Pillow==10.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.3
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
